xdiff: include new xdiff from git
Update to the xdiff used in git v2.35.0, with minor changes to avoid
the sort of warnings that we care about (signed/unsigned mismatch,
unused, etc.)

Any git-specific abstraction bits are now redefined for our use in
`git-xdiff.h`.  It is a (wildly optimistic) hope that we can use that
indirection layer to standardize on a shared xdiff implementation.

Also disable warnings for xdiff because we're now regressing there.